make NonEmptyStreamSuite 2.12- specific
